+++ src/CleanRegex/Internal/Match/Details/Groups/GroupArrayKey.php
@@ -1,0 +1,9 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Match\Details\Groups;
+
+interface GroupArrayKey
+{
+    public function applies($nameOrIndex): bool;
+
+    public function key($nameOrIndex);
+}