+++ src/CleanRegex/Internal/Pcre/DeprecatedMatchDetail.php
@@ -5,6 +5,7 @@ use TRegx\CleanRegex\Internal\Match\Details\Group\GroupFactoryStrategy;
 use TRegx\CleanRegex\Internal\Match\Details\Group\MatchGroupFactoryStrategy;
 use TRegx\CleanRegex\Internal\Pcre\Legacy\IRawMatchOffset;
 use TRegx\CleanRegex\Internal\Pcre\Legacy\MatchAllFactory;
+use TRegx\CleanRegex\Internal\Pcre\Legacy\Prime\Prime;
 use TRegx\CleanRegex\Internal\Pcre\Signatures\PerformanceSignatures;
 use TRegx\CleanRegex\Internal\Subject;
 use TRegx\CleanRegex\Match\Details\MatchDetail;
@@ -17,12 +18,16 @@ class DeprecatedMatchDetail
     /**
      * @deprecated
      */
-    public static function create(Subject              $subject, int $index,
-                                  IRawMatchOffset      $match, MatchAllFactory $allFactory,
+    public static function create(Subject              $subject,
+                                  int                  $index,
+                                  IRawMatchOffset      $match,
+                                  MatchAllFactory      $allFactory,
+                                  Prime                $prime,
                                   GroupFactoryStrategy $strategy = null): MatchDetail
     {
-        return new MatchDetail($subject, $index, $match, $match, $match, $match, $allFactory,
+        return new MatchDetail($subject, $index, $match, $match, $match, $allFactory,
             $strategy ?? new MatchGroupFactoryStrategy(),
-            new PerformanceSignatures($match, $match));
+            new PerformanceSignatures($match, $match),
+            $prime);
     }
 }