+++ src/CleanRegex/Internal/Pcre/Legacy/Prime/MatchesFirstPrime.php
@@ -1,0 +1,27 @@
+<?php
+namespace TRegx\CleanRegex\Internal\Pcre\Legacy\Prime;
+
+use TRegx\CleanRegex\Internal\Model\Entry;
+use TRegx\CleanRegex\Internal\Pcre\Legacy\RawMatchesOffset;
+use TRegx\CleanRegex\Internal\Pcre\Legacy\UsedForGroup;
+
+class MatchesFirstPrime implements Prime
+{
+    /** @var RawMatchesOffset */
+    private $matches;
+
+    public function __construct(RawMatchesOffset $matches)
+    {
+        $this->matches = $matches;
+    }
+
+    public function firstUsedForGroup(): UsedForGroup
+    {
+        return new MatchesFirstUsedForGroup($this->matches);
+    }
+
+    public function firstEntry(): Entry
+    {
+        return new MatchesFirstEntry($this->matches);
+    }
+}