+++ src/CleanRegex/Replace/By/LazyDetail.php
@@ -7,8 +7,6 @@ use TRegx\CleanRegex\Internal\Subject;
 use TRegx\CleanRegex\Match\Details\Detail;
 use TRegx\CleanRegex\Match\Details\DuplicateName;
 use TRegx\CleanRegex\Match\Details\Group\Group;
-use TRegx\CleanRegex\Match\Details\Groups\IndexedGroups;
-use TRegx\CleanRegex\Match\Details\Groups\NamedGroups;
 
 class LazyDetail implements Detail
 {
@@ -119,12 +117,18 @@ class LazyDetail implements Detail
         return $this->detail()->usingDuplicateName();
     }
 
-    public function groups(): IndexedGroups
+    /**
+     * @return Group[]
+     */
+    public function groups(): array
     {
         return $this->detail()->groups();
     }
 
-    public function namedGroups(): NamedGroups
+    /**
+     * @return Group[]
+     */
+    public function namedGroups(): array
     {
         return $this->detail()->namedGroups();
     }